+++ src/mig/api/context.go
@@ -48,7 +48,7 @@ import (
 // database and logging. It also contains some statistics.
 // Context is intended as a single structure that can be passed around easily.
 type Context struct {
-	OpID     uint64 // ID of the current operation, used for tracking
+	OpID     float64 // ID of the current operation, used for tracking
 	Channels struct {
 		Log chan mig.Log
 	}
